requirements-desktop.txt
fastapi>=0.104.0
uvicorn>=0.24.0
psutil>=5.9.0
numpy>=1.24.0
cryptography>=41.0.0
pydantic>=2.0.0
python-multipart>=0.0.6
aiofiles>=23.0.0
PyQt6>=6.6.0
pyinstaller>=6.0.0
pillow>=10.0.0
requests>=2.31.0
